Support for sk-ecdsa-sha2-nistp256 and sk-ssh-ed25519 keys, FIDO (#698)
Notes:
Add support for sk-ecdsa-sha2-nistp256@openssh.com and sk-ssh-ed25519@openssh.com key exchange for FIDO auth using the OpenSSL backend. Stub API for other backends.

Credit:
Michael Buckley

+++ tests/test_keyboard_interactive_auth_fails_with_wrong_response.c
@@ -19,8 +19,8 @@ static void kbd_callback(const char *name, int name_len,
     fprintf(stdout, "Kb-int name: %.*s\n", name_len, name);
     fprintf(stdout, "Kb-int instruction: %.*s\n", instruct_len, instruct);
     for(i = 0; i < num_prompts; ++i) {
-        fprintf(stdout, "Kb-int prompt %d: %.*s\n", i, prompts[i].length,
-                prompts[i].text);
+        fprintf(stdout, "Kb-int prompt %d: %.*s\n", i,
+                (int)prompts[i].length, prompts[i].text);
     }
 
     if(num_prompts == 1) {
